Move Backtracking reftypes analysis impl in analysis_main.
(code-motion only)
